Updated script and ran paper-ready evals

+++ source/run_models.py
@@ -1,35 +1,40 @@
 import json
 import os
 import asyncio
+import re
 import argparse
+import traceback
 from datetime import datetime
 
 def parse_arguments():
     parser = argparse.ArgumentParser(description='')
+    # General args
     parser.add_argument('-m', '--model', type=str, help='model name')
     parser.add_argument('-p', '--prompt', type=str, help='prompt name')
-    parser.add_argument('--context', type=str, help='new, day, sum')
-    parser.add_argument('-c', '--case', type=str, help='If ALL, run all cases; if a case number like 3-4-1, run that case; if a case number followed by a "+" like 3-4-1+, run that case and all cases after it.')
-    parser.add_argument('-n', '--no_description', action='store_true')
+    parser.add_argument('--context', type=str, help='sofar, today, sum')
+    parser.add_argument('--case', type=str, default="ALL", help='If ALL, run all cases; if a case number like 3-4-1, run that case; if a case number followed by a "+" like 3-4-1+, run that case and all cases after it.')
+    parser.add_argument('--no_description', action='store_true')
+
+    # Evaluation args
+    parser.add_argument('-a', '--all', action='store_true', help='Evaluate all existing models')
     return parser
 
 def get_output_dir(MODEL, PROMPT, CONTEXT, CASE, NO_DESCRIPTION):
-    output_dir = f'../output/{MODEL.split("/")[-1]}_{PROMPT}'
+    output_dir = f'../output/{MODEL.split("/")[-1]}_prompt_{PROMPT}'
     if CONTEXT is not None:
         output_dir += f"_context_{CONTEXT}"
     if NO_DESCRIPTION:
-        output_dir += "_no_desc"    
+        output_dir += "_desc_none"    
     if CASE != "ALL":
         output_dir += f"_case_{CASE}"
     return output_dir
 
-def get_fnames(data_dir, output_dir, CASE, eval=False):
+def get_fnames(data_dir, output_dir, CASE, eval=False, verbose=True):
     """Return list of .json files"""
     all_fnames = sorted([
         fname for fname 
         in os.listdir(data_dir) 
         if fname.endswith('.json')
-        if int((fname.split("_")[0]).split("-")[-1]) % 2 == 0
         if not fname.startswith(('7-', '8-'))  # skip test split
     ])
     fnames = []
@@ -44,7 +49,7 @@ def get_fnames(data_dir, output_dir, CASE, eval=False):
                     fnames = [fname]
                 break
 
-    print(f"<get_fnames> Found {len(fnames)} cases")
+    if verbose: print(f"<get_fnames> Found {len(fnames)} total cases")
 
     if not eval:
         fnames_to_check = fnames.copy()
@@ -53,7 +58,7 @@ def get_fnames(data_dir, output_dir, CASE, eval=False):
                 # print(f"Skipping existing {fname.split('.')[0] + '.jsonl'}")
                 fnames.remove(fname)
 
-    print(f"<get_fnames> Running {len(fnames)} new cases")
+    if verbose:print(f"<get_fnames> Running {len(fnames)} new cases")
     return fnames
 
 def build_prompt_prefix_suffix(prompt_arg):
@@ -81,12 +86,14 @@ def parse_json(file_path):
         data = json.load(file)
         characters = []
         evidences = []
+        prev_context = re.sub(r'\n+', ' ', data.get("previousContext", ""))  # Remove newlines
+
         for character in data.get('characters', {}):
             characters.append(character)
         for evidence in data.get('evidences', {}):
             evidences.append(evidence)
         turns = []
-        for turn in data["turns"]:
+        for turn in data.get("turns", []):
             if turn["noPresent"]:
                 continue
             testimonies = []
@@ -96,32 +103,56 @@ def parse_json(file_path):
                 'characters': characters,
                 'evidences': evidences,
                 'testimonies': testimonies,
-                'new_context': turn['newContext']
+                'new_context': re.sub(r'\n+', ' ', turn['newContext'])
             }
             summarized_context = ""
             if 'summarizedContext' in turn: summarized_context = turn['summarizedContext']
             elif 'summarized_context' in turn: summarized_context = turn['summarized_context']
             turn_dict['summarized_context'] = summarized_context
             turns.append(turn_dict)
-        return turns, data["previousContext"]
+        return turns, prev_context
 
-def build_prompt(turns, prev_context, PROMPT_PREFIX, PROMPT_SUFFIX, CONTEXT, NO_DESCRIPTION):
+def build_prompt(
+    turns, 
+    prev_context, 
+    PROMPT_PREFIX, 
+    PROMPT_SUFFIX, 
+    CONTEXT, 
+    NO_DESCRIPTION, 
+    MODEL
+):
     prompts = []
     context_sofar = ""
     for turn in turns:
         context_is_added = False
-        new_context = turn['new_context']
+        new_context = turn['new_context']  
+        new_context = re.sub(r'\n+', ' ', new_context)  # Remove newlines
         context_sofar += new_context
         if CONTEXT is None:
             prompt = ""
         else:
             prompt = "Story:\n"
-            if CONTEXT == "new":
-                prompt += prev_context + "\n" + new_context + "\n"
-            elif CONTEXT == "day":
-                prompt += prev_context + "\n" + context_sofar + "\n"
+            full_context = "" 
+            if CONTEXT == "today":
+                full_context += prev_context + "\n" + new_context + "\n"
+            elif CONTEXT == "sofar":
+                full_context += prev_context + "\n" + context_sofar + "\n"
             elif CONTEXT == "sum":
-                prompt = "Summarized Context:\n" + turn['summarized_context'] + "\n"
+                full_context += turn['summarized_context'] + "\n"
+            context_size = len(full_context)  # Count characters, not tokens
+
+            # Truncate context for specific models
+            max_context_size = -1
+            if "deepseek-reasoner" in MODEL or "deepseek-chat" in MODEL:  # Roughly 66000 tokens for deepseek
+                max_context_size = 230000
+            elif "70b" in MODEL:
+                max_context_size = 80000
+            if max_context_size != -1:
+                start_idx = context_size - max_context_size
+                full_context = "..." + full_context[start_idx:]
+
+            prompt += full_context
+
         character_counter = 0
         prompt += "Characters:\n"
         for character in turn['characters']:
@@ -185,62 +216,92 @@ def get_json_answer(multiline_string):
     target = lines[-1]
     try:
         json_answer = json.loads(target)
+        cot = "\n".join(lines[:-1])
     except json.JSONDecodeError:
         try:
             target = lines[-2]
             json_answer = json.loads(target)
+            cot = "\n".join(lines[:-2])
         except json.JSONDecodeError:
             # print(f"Error parsing JSON: {multiline_string[:-1]}")
             json_answer = {}
-    
-    return json_answer
+            cot = ""
+    return json_answer, cot
 
 def run_model(prompts, client, client_name):
+    has_error = False
     answer_jsons = []
-    full_responses = []
+    cots = []
     for prompt in prompts:
         #print(prompt)
-        if type(client).__name__ == "Kani":  # Use kani api
-            async def run_model():
-                response = await client.chat_round_str(prompt, temperature=0.6)
-                #print(response)
-                return response
+        try:
+            cot = ""
+            if type(client).__name__ == "Kani":  # Use kani api
+                async def run_async_model():
+                    response = await client.chat_round_str(prompt, temperature=0.6)
+                    #print(response)
+                    return response
 
-            full_answer = asyncio.run(run_model())
+                full_answer = asyncio.run(run_async_model())
 
-        elif type(client).__name__ == "OpenAI":  # Use openai api
-            response = client.chat.completions.create(
-                model=client_name,
-                messages=[
+            elif type(client).__name__ == "OpenAI":  # Use openai api
+                response = client.chat.completions.create(
+                    model=client_name,
+                    messages=[
                     {"role": "system", "content": "You are a helpful assistant"},
                     {"role": "user", "content": prompt},
                 ],
-                stream=False
-            )
+                    stream=False
+                )
+                full_answer = response.choices[0].message.content
 
-            full_answer = response.choices[0].message.content
-            try: 
-                cot = response.choices[0].message.reasoning_content
-                print(f"COT returned for {client_name}")
-                full_answer = cot + "\n\n" + full_answer
-            except Exception as e:
-                print(f"When trying to get COT for {client_name}: {e}")
-                print(f"No COT for {client_name}")
+                # Get COT
+                try: 
+                    cot = response.choices[0].message.reasoning_content
+                    print(f"<run_model> COT returned for {client_name}")
+                    full_answer = cot + "\n\n" + full_answer
+                except Exception as e:
+                    print(f"<run_model> When trying to get COT for {client_name}: {e}")
+                    print(f"<run_model> No COT for {client_name}")
 
-        else:
-            raise ValueError(f"Unknown client: {client}")
+            else:
+                raise ValueError(f"<run_model> Unknown client: {client}")
 
-        answer_json = get_json_answer(full_answer)
+            answer_json, parsed_cot = get_json_answer(full_answer)
+
+            if answer_json == {}:
+                has_error = True
+
+            if cot == "":  # Only when model does not return its COT field
+                cot = parsed_cot
+                
+        except Exception as e:  # Handle errors, such as rate limit, context window, etc.
+            print(f"<run_model> {traceback.format_exc()}")
+            answer_json, cot = {}, ""
+            has_error = True
+
         answer_jsons.append(answer_json)
-        full_responses.append(full_answer)
+        cots.append(cot)
 
-    return answer_jsons, full_responses
+    return answer_jsons, cots, has_error
 
 def load_model(model):
-    if "/" in model:  # a huggingface model
+    if any(m_name in model for m_name in ["llama", "deepseek-R1-70b", "deepseek-R1-32b", "deepseek-R1-8b"]) or \
+    "/" in model:  # a huggingface model
         from kani import Kani
         from kani.engines.huggingface import HuggingEngine
         import torch
+
+        if model == "llama-3.1-70b":
+            model = "meta-llama/Llama-3.1-70B-Instruct"
+        elif model == "llama-3.1-8b":
+            model = "meta-llama/Llama-3.1-8B-Instruct"
+        elif model == "deepseek-R1-70b":
+            model = "deepseek-ai/DeepSeek-R1-Distill-Llama-70B"
+        elif model == "deepseek-R1-32b":
+            model = "deepseek-ai/DeepSeek-R1-Distill-Qwen-32B"
+        elif model == "deepseek-R1-8b":
+            model = "deepseek-ai/DeepSeek-R1-Distill-Llama-8B"
 
         torch.cuda.empty_cache()
         engine = HuggingEngine(
@@ -259,14 +320,30 @@ def load_model(model):
         load_dotenv("../.env")
 
         model_key = model
-        if model in ["gpt-4o-mini", "gpt-4o", "o3-mini"]:
+
+        if any(m_name in model for m_name in ["gpt", "o3", "o4"]):
             model_key = "openai"
+            if model == "gpt-4o":
+                model = "gpt-4o-2024-08-06"
+            elif model == "gpt-4o-mini":
+                model = "gpt-4o-2024-07-18"
+            elif model == "o3-mini":
+                model = "o3-mini-2025-01-31"
+            elif model == "o4-mini":
+                model = "o4-mini-2025-04-16"
+            elif model == "gpt-4.1":
+                model = "gpt-4.1-2025-04-14"
+            elif model == "gpt-4.1-mini":
+                model = "gpt-4.1-mini-2025-04-14"
+
+        elif "deepseek" in model:  # deepseek-reasoner (R1), deepseek-chat (V3)
+            model_key = "deepseek"
 
         auth = {
             "deepseek": {
                 "api_key": os.getenv("DEEPSEEK_API_KEY"),
                 "base_url": "https://api.deepseek.com",
-                "name": "deepseek-reasoner"
+                "name": model
             },
             "openai": {
                 "api_key": os.getenv("OPENAI_API_KEY"),
@@ -292,10 +369,14 @@ def create_batch(fnames, MODEL, PROMPT, CONTEXT, NO_DESCRIPTION, data_dir):
     max_token_key = "max_tokens" if "gpt" in MODEL else "max_completion_tokens"
     max_token_val = 1000 if "gpt" in MODEL else 5000
     batch = []
+    skip_count = 0
     for fname in fnames:
         turns, prev_context = parse_json(os.path.join(data_dir, fname))
+        if turns == []:
+            skip_count += 1
+            continue
         PROMPT_PREFIX, PROMPT_SUFFIX = build_prompt_prefix_suffix(PROMPT)
-        prompts = build_prompt(turns, prev_context, PROMPT_PREFIX, PROMPT_SUFFIX, CONTEXT, NO_DESCRIPTION)
+        prompts = build_prompt(turns, prev_context, PROMPT_PREFIX, PROMPT_SUFFIX, CONTEXT, NO_DESCRIPTION, MODEL)
         # print(prompts)
         for i, prompt in enumerate(prompts):
             request = {
@@ -314,6 +395,8 @@ def create_batch(fnames, MODEL, PROMPT, CONTEXT, NO_DESCRIPTION, data_dir):
             batch.append(request)
     
     assert len(batch) > 0, "Must have at least 1 batch item"
+    print(f"Skipped {skip_count} cases")
+
     return batch
 
 def submit_batch_job(jsonl_path, client, output_dir):
@@ -342,24 +425,74 @@ def submit_batch_job(jsonl_path, client, output_dir):
 
 def run_batch_job(fnames, MODEL, PROMPT, CONTEXT, NO_DESCRIPTION, client, output_dir, data_dir):
     jsonl_path = os.path.join(output_dir, "batchinput.jsonl")
-    batch = create_batch(fnames, MODEL, PROMPT, CONTEXT, NO_DESCRIPTION, data_dir)
-    with open(jsonl_path, "w") as f:
-        for request in batch:
-            f.write(json.dumps(request, ensure_ascii=False) + "\n")
+
+    # If exists, run the incomplete batch job instead
+    if os.path.exists(jsonl_path):
+        # Create duplicate file
+        dup_count = 1
+        new_jsonl_path = jsonl_path
+        while os.path.exists(new_jsonl_path):
+            new_jsonl_path = os.path.join(output_dir, f"batchinput_{dup_count}.jsonl")
+            dup_count += 1
+        print(f"<run_batch_job> Creating duplicate batch input file: {os.path.basename(new_jsonl_path)}")
+
+        # Filter incomplete input
+        with open(jsonl_path, "r") as f:
+            input_data = [json.loads(line) for line in f]
+        output_data = []
+        output_files = sorted([
+            os.path.join(output_dir, output_path) 
+            for output_path in os.listdir(output_dir) 
+            if output_path.startswith("batchoutput")
+        ])  # Guaranteed to be mutually exclusive
+        for output_file in output_files:
+            with open(output_file, "r") as f:
+                output_data += [json.loads(line) for line in f]
+        output_ids = set([item["custom_id"] for item in output_data])
+        new_input_data = []
+        for input_item in input_data:
+            if input_item["custom_id"] not in output_ids:
+                new_input_data.append(input_item)
+        print(f"<run_batch_job> Found {len(new_input_data)} incomplete cases")
+
+        # Create new batch input file
+        with open(new_jsonl_path, "w") as f:
+            for request in new_input_data:
+                f.write(json.dumps(request, ensure_ascii=False) + "\n")
+        jsonl_path = new_jsonl_path
+
+    else:
+        batch = create_batch(fnames, MODEL, PROMPT, CONTEXT, NO_DESCRIPTION, data_dir)
+        with open(jsonl_path, "w") as f:
+            for request in batch:
+                f.write(json.dumps(request, ensure_ascii=False) + "\n")
     
     batch_job_id = submit_batch_job(jsonl_path, client, output_dir)
     return batch_job_id
 
 def run_job(fnames, MODEL, PROMPT, CONTEXT, NO_DESCRIPTION, client, client_name, output_dir, data_dir):
+    error_count = 0
+    skip_count = 0
     for fname in fnames:
         # Parse and build prompt
+        if error_count > 5:
+            print(f"<run_job> Terminating due to {error_count}+ json parsing errors")
+            break
+
         print(fname)
         turns, context = parse_json(os.path.join(data_dir, fname))
+        if turns == []:  # Skip cases with no turns
+            skip_count += 1
+            continue
         PROMPT_PREFIX, PROMPT_SUFFIX = build_prompt_prefix_suffix(PROMPT)
-        prompts = build_prompt(turns, context, PROMPT_PREFIX, PROMPT_SUFFIX, CONTEXT, NO_DESCRIPTION)
+        prompts = build_prompt(turns, context, PROMPT_PREFIX, PROMPT_SUFFIX, CONTEXT, NO_DESCRIPTION, MODEL)
 
         # Answer
-        answer_jsons, full_responses = run_model(prompts, client, client_name)
+        answer_jsons, cots, has_error = run_model(prompts, client, client_name)
+        if has_error:
+            error_count += 1
+            print(f"<run_job> Error when running the model for {fname}")
+            continue  # Skip cases with errors
         for answer_json in answer_jsons:
             print(answer_json)
 
@@ -369,14 +502,16 @@ def run_job(fnames, MODEL, PROMPT, CONTEXT, NO_DESCRIPTION, client, client_name,
                 file.write(json.dumps(answer_json) + "\n")
         with open(os.path.join(output_dir, fname.split('.')[0] + '_outputs.json'), 'w') as file:
             json_response = []
-            for idx, response in enumerate(full_responses):
-                json_response.append({
+            for idx, (answer_json, cot) in enumerate(zip(answer_jsons, cots)):
+                json_response.append({ 
                     "idx": idx,
                     "prompt": prompts[idx],
-                    "response": response,
-                    "response_json": answer_jsons[idx]
+                    "response_json": answer_json,
+                    "cot": cot
                 })
             file.write(json.dumps(json_response, indent=2))
+    
+    print(f"Skipped {skip_count} cases")
 
 if __name__ == "__main__":
     parser = parse_arguments()
@@ -398,8 +533,8 @@ if __name__ == "__main__":
         json.dump({
             'model': MODEL,
             'prompt': PROMPT,
-            'context': CONTEXT,
-            'case': CASE,
+            'context': "none" if CONTEXT is None else CONTEXT,
+            'case': CASE if CASE != "ALL" else "all",
             'no_description': NO_DESCRIPTION,
             'timestamp': timestamp
         }, file, indent=2)
@@ -410,7 +545,7 @@ if __name__ == "__main__":
     fnames = get_fnames(data_dir, output_dir, CASE)
 
     # Run cases
-    if MODEL in ["gpt-4o-mini", "gpt-4o", "o3-mini"]:
+    if any(name in MODEL for name in ["o3", "o4", "gpt"]):
         run_batch_job(fnames, MODEL, PROMPT, CONTEXT, NO_DESCRIPTION, client, output_dir, data_dir)
     else:
         run_job(fnames, MODEL, PROMPT, CONTEXT, NO_DESCRIPTION, client, client_name, output_dir, data_dir)
